fix(polish): loop-5 — 99 clippy warnings + 5 CLI guard tests + cargo-audit
Loop-5 scope-expanded beyond beads to workspace-wide lint/test/audit
cleanliness. Two parallel agents drove the tail-end residuals to zero.

**Clippy cleanup (99 → 0)**

- 87 warnings auto-fixed via `cargo clippy --fix` across ~37 files
  (div_ceil, is_multiple_of, to_vec, redundant closures, collapsible
  if, split_once, field_reassign_with_default, etc.)
- 12 manually fixed:
  - pcloud-fs/tests/chunked_upload_write_multi_gib.rs: #[allow(dead_code)]
    on test-only SessionState (fields retained for debug output)
  - pcloud-fs/tests/fuse_adapter_unit.rs: MockDirEntry<'a> type alias
    resolves type_complexity
  - pcloud-crypto/src/lib.rs: struct-literal update syntax for
    CryptoShell::default() assign-then-mutate pattern
  - pcloud-backends/src/transfer_backend.rs: blank line closes rustdoc
    list before new paragraph
  - pcloud-backends/tests/crypto_share_rsa_e2e.rs: prose "+" → "and"
    to drop false list-continuation warning

**Test suite fixes (5 CLI argv-guard tests)**

pcloud-cli/src/app.rs argv-secret guard was missing allowed-flags
entries. 5 tests passed secrets on argv without `--allow-argv-password`
and were crashing the harness via `process::exit(2)`:
- app.rs:981-988: extended `allowed_flags_for` to include
  --allow-argv-password for SubmitTwoFactorCode + SubmitRecoveryCode,
  added entries for SubmitAuthToken + SubmitCryptoPassword
- 5 tests updated to pass --allow-argv-password:
  - submit_auth_parses_explicit_token
  - submit_tfa_parses_explicit_code_and_flag
  - submit_recovery_marks_recovery_code_path
  - two_token_crypto_start_password_shifts_correctly
  - legacy_tfa_alias_routes_to_submit_tfa

**cargo-audit / cargo-deny**

- RUSTSEC-2023-0071 (rsa Marvin timing sidechannel, medium, no
  upstream patch) → new bead pcloud-rs-ktj (P2) filed + added to
  deny.toml ignore list with review date 2026-07-15. Rationale:
  upstream has no patched version; impact limited to the OAEP
  unwrap path in share_temppass/share_rsa.
- `paste` + `rustls-pemfile` unmaintained warnings already triaged
  in deny.toml or transitively pinned — no action.

**Workspace final state**
- cargo check --workspace: 0 rustc warnings
- cargo clippy --workspace --all-targets --no-deps: 0 clippy warnings
- cargo test --workspace --lib --no-fail-fast: 1557 passed, 0 failed,
  3 ignored across 33 crates
- cargo test --workspace --tests: 2217 passed, 0 failed, 52 ignored
- cargo deny check: advisories ok, bans ok, licenses ok, sources ok
- 1 open bead: pcloud-rs-ktj (deferred — no upstream patch available)

+++ crates/pcloud-daemon/src/sync_loop_runtime.rs
@@ -631,7 +631,7 @@ impl SyncLoopRuntime for RealSyncLoopRuntime {
                                 // is still recognised as non-stalled
                                 // via its byte counter.
                                 self.stall_detector
-                                    .observe_bytes(path, written as u64);
+                                    .observe_bytes(path, written);
                                 if self.engine.mark_transfer_completed(path) {
                                     completed += 1;
                                     // P2-b (H2): durable ack — remove
